mgr/nfs: rename fs dir to export
